# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Personal Health Companion API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /symptoms:
    get:
      operationId: getSymptomLogs
      tags: [symptoms]
      summary: Get symptom log history
      parameters:
        - name: x-user-id
          in: header
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Symptom logs
          content:
            application/json:
              schema:
                type: object
                properties:
                  logs:
                    type: array
                    items:
                      $ref: "#/components/schemas/SymptomLog"
                required: [logs]
    post:
      operationId: analyzeSymptoms
      tags: [symptoms]
      summary: Log symptoms and get AI analysis
      parameters:
        - name: x-user-id
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SymptomInput"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SymptomLog"
        "402":
          description: Insufficient credits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: AI service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
...
    SymptomInput:
      type: object
      properties:
        inputText:
          type: string
          minLength: 3
      required: [inputText]
    SymptomLog:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        inputText:
          type: string
        riskScore:
          type: integer
          minimum: 0
          maximum: 100
        tier:
          type: string
          enum: [home_care, monitor, see_doctor_24h, emergency]
        aiReason:
          type: string
        advice:
          type: array
          items:
            type: string
        isEmergency:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, userId, inputText, riskScore, tier, aiReason, advice, isEmergency, createdAt]
...