# lib/api-zod/src/generated/types/symptomLog.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AI Personal Health Companion API
 * OpenAPI spec version: 0.1.0
 */
import type { SymptomLogTier } from './symptomLogTier';

export interface SymptomLog {
  id: string;
  userId: string;
  inputText: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  riskScore: number;
  tier: SymptomLogTier;
  aiReason: string;
  advice: string[];
  isEmergency: boolean;
  createdAt: Date;
}